3 ошибки: в методах conver, showNextQuestionOrResult, и по протоколу в QuestionFactory
